+++ database/migrations/2026_05_09_161300_create_event_assignment_groups_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignIdFor(Event::class)->constrained()->cascadeOnDelete();
             $table->string('name')->default('New group');
             $table->longText('notes')->nullable();
-            $table->unsignedTinyInteger('sort_order')->default(0);
+            $table->integer('sort_order')->default(0);
             $table->timestamps();
         });
