+++ database/migrations/2026_08_21_182433_make_event_assignments_sort_order_not_nullable.php
@@ -9,14 +9,14 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('event_assignments', function (Blueprint $table) {
-            $table->unsignedSmallInteger('sort_order')->nullable(false)->change();
+            $table->integer('sort_order')->nullable(false)->change();
         });
     }
 
     public function down(): void
     {
         Schema::table('event_assignments', function (Blueprint $table) {
-            $table->unsignedSmallInteger('sort_order')->nullable()->change();
+            $table->integer('sort_order')->nullable()->change();
         });
     }
 };
